Unifies environment of mobile robot
